Feature/video speedup

+++ components/box-emu-hal/include/statistics.hpp
@@ -5,13 +5,13 @@
 
 #include "format.hpp"
 
-void update_frame_time(float frame_time);
+void update_frame_time(uint64_t frame_time);
 void reset_frame_time();
 
 float get_fps();
-float get_frame_time();
-float get_frame_time_max();
-float get_frame_time_min();
+uint64_t get_frame_time();
+uint64_t get_frame_time_max();
+uint64_t get_frame_time_min();
 float get_frame_time_avg();
 
 void print_statistics();
